+++ src/main/java/com/buuz135/industrial/tile/agriculture/CropRecolectorTile.java
@@ -76,7 +76,7 @@ public class CropRecolectorTile extends WorkingAreaElectricMachine {
         List<BlockPos> blockPos = BlockUtils.getBlockPosInAABB(getWorkingArea());
         boolean shouldPointerIncrease = true;
         boolean didWork = false;
-        if (pointer < blockPos.size()) {
+        if (pointer < blockPos.size() && BlockUtils.canBlockBeBroken(this.world, blockPos.get(pointer))) {
             BlockPos pos = blockPos.get(pointer);
             IBlockState state = this.world.getBlockState(blockPos.get(pointer));
             Optional<PlantRecollectable> recollectable = IFRegistries.PLANT_RECOLLECTABLES_REGISTRY.getValues().stream().sorted(Comparator.comparingInt(PlantRecollectable::getPriority)).filter(iPlantRecollectable -> iPlantRecollectable.canBeHarvested(this.world, pos, state)).findFirst();