+++ src/main/java/com/buuz135/industrial/tile/agriculture/CropSowerTile.java
@@ -10,10 +10,12 @@ import net.minecraft.item.EnumDyeColor;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
+import net.minecraftforge.common.ForgeHooks;
 import net.minecraftforge.common.IPlantable;
 import net.minecraftforge.common.util.FakePlayer;
 import net.ndrei.teslacorelib.gui.BasicTeslaGuiContainer;
@@ -111,10 +113,9 @@ public class CropSowerTile extends WorkingAreaElectricMachine {
                     if (!ItemStackUtils.isStackOreDict(stack, "treeSapling") && (this.world.getBlockState(pos.offset(EnumFacing.DOWN)).getBlock().equals(Blocks.DIRT) || this.world.getBlockState(pos.offset(EnumFacing.DOWN)).getBlock().equals(Blocks.GRASS))) {
                         this.world.setBlockState(pos.offset(EnumFacing.DOWN), Blocks.FARMLAND.getDefaultState());
                     }
-                    Item seeds = stack.getItem();
                     player.setHeldItem(EnumHand.MAIN_HAND, stack);
-                    seeds.onItemUse(player, world, pos.offset(EnumFacing.DOWN), EnumHand.MAIN_HAND, EnumFacing.UP, 0, 0, 0);
-                    return 1;
+                    EnumActionResult result = ForgeHooks.onPlaceItemIntoWorld(stack, player, world, pos, EnumFacing.UP, 0,0,0, EnumHand.MAIN_HAND);
+                    return result == EnumActionResult.SUCCESS ? 1 : 0;
                 }
             }
         } else {