+++ src/main/java/com/buuz135/industrial/utils/BlockUtils.java
@@ -1,5 +1,6 @@
 package com.buuz135.industrial.utils;
 
+import com.buuz135.industrial.IndustrialForegoing;
 import net.minecraft.block.BlockLeaves;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.renderer.*;
@@ -9,9 +10,13 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
+import net.minecraft.util.EnumHand;
 import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.common.util.BlockSnapshot;
+import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.oredict.OreDictionary;
 import org.lwjgl.opengl.GL11;
 
@@ -52,6 +57,13 @@ public class BlockUtils {
     public static boolean isLeaves(World world, BlockPos pos) {
         return world.getBlockState(pos).getBlock() instanceof BlockLeaves || isBlockOreDict(world, pos, "treeLeaves");
     }
+
+    public static boolean canBlockBeBroken(World world, BlockPos pos){
+        BlockEvent.BreakEvent event = new BlockEvent.BreakEvent(world, pos, world.getBlockState(pos), IndustrialForegoing.getFakePlayer(world));
+        MinecraftForge.EVENT_BUS.post(event);
+        return !event.isCanceled();
+    }
+
 
     public static void renderLaserBeam(TileEntity tile, double x, double y, double z, EnumFacing direction, float partialTicks, int length) {
         Tessellator tess = Tessellator.getInstance();